requirements.txt
openai
python-dotenv
pandas
gspread
gspread_dataframe
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
PyMuPDF
easyocr
Pillow
numpy
streamlit
streamlit-authenticator